requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
discord==2.0.0
discord.py==2.0.1
emoji==1.6.3
frozenlist==1.3.3
idna==3.4
joblib==1.2.0
multidict==6.0.2
nltk==3.7
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
regex==2022.10.31
requests==2.28.1
six==1.16.0
spotipy==2.21.0
text2emotion==0.0.5
tqdm==4.64.1
urllib3==1.26.12
wrapt==1.14.1
yarl==1.8.1
